refactor: DayComponent renamed to DayComponentView, modularized dayData.map return into DayComponent Component, minor functionality
